Added: 1) Init state section to Add_Data_Screen 2) Initialized clearing of fields on Add_Data_Screen 3) Used return value screen_pop for amount update in Add_Data_Screen. 4) Clear field values triggered on back button from Add_Data_Screen. 5) Clear field values triggered on back button from Loan_Approve_Screen. 6) Conditons for monthly and loan approve screen functions. 7) Clear state functions collection made to single function. 8) Created another common variable file for multiselect_Screen 9) Created file for utilities in mixed collection filed. 10) Popup alerts created for Add_Data_Screen. 11) Popup alerts created for Loan_Approve_Screen.
Changes:
1) Changed all values pulled from db to double for double data type variable.
2) Changed some variable data type; int, var to double, string.
3) Changed variable file location.
4) Renamed some variables names.
5) Renamed some functions and utilities names for transparency.
6) Changed keyboard type to text from number for comments in Loan_Approve_Screen.
7) Removed unused import files.

+++ lib/functions/firestoreFunctions/get_total_fund_recieved_wig.dart
@@ -12,7 +12,7 @@ Future<void> getTotalFundRecievedFromDb() async {
   DocumentSnapshot docTotalSnapshot = await docTotal.get();
 
   totalFundRecievedPulledFromDB =
-      docTotalSnapshot['balance_fund_from_true_count_monthly'];
+      (docTotalSnapshot['balance_fund_from_true_count_monthly']).toDouble();
 
   // total expected fund calculated here since this function is executing last in initialization
   totalFundExpectedCalculatedLocal = balanceFundTotalPulledFromDB +
